+++ packages/component-generator/src/createComponents.tsx
@@ -3,9 +3,9 @@ import { isValidElementType } from 'react-is';
 import isPlainObject from 'lodash.isplainobject';
 import isEmpty from 'lodash.isempty';
 import { createStyledComponent, createFC } from './createStyledComponent';
-
 import { Settings, BaseConfig, Variant, ComponentsRecord } from './typings';
 
+/* eslint-disable @typescript-eslint/no-explicit-any */
 export function createComponents<
   Config = any,
   ThemeType = any,
@@ -17,7 +17,7 @@ export function createComponents<
     | React.FC<Props>,
   settings: Settings<Element>
 ): ComponentsRecord<Config, Props, ThemeType> {
-  const dict = createBaseComponent<Config, ThemeType, Props>(base);
+  const dict = createBaseComponent<Config, ThemeType, Props, Element>(base);
   const isConfigBase = dict.hasOwnProperty('Base');
 
   if ((isValidElementType(base) || !isEmpty(base)) && !isEmpty(settings)) {
@@ -26,8 +26,11 @@ export function createComponents<
 
       if (hasKey(settings, prop)) {
         dict[prop] = isConfigBase
-          ? generateComponent<ThemeType, Props, Element>(base, setting)
-          : createFC<Props>(base, setting);
+          ? generateComponent<ThemeType, Props, Element>(
+              base as BaseConfig<Props, AllHTMLAttributes<Element>, ThemeType>,
+              setting as Settings<Element>
+            )
+          : createFC<Props>(base as React.FC<Props>, setting);
         dict[prop].displayName = prop;
       }
       return dict;
@@ -42,7 +45,11 @@ export function createBaseComponent<
   ThemeType = any,
   Props = Record<string, unknown>,
   Element = HTMLDivElement
->(base): ComponentsRecord<Config, Props, ThemeType> {
+>(
+  base:
+    | BaseConfig<Props, AllHTMLAttributes<Element>, ThemeType>
+    | React.FC<Props>
+): ComponentsRecord<Config, Props, ThemeType> {
   const dict = {} as ComponentsRecord<Config, Props, ThemeType>;
 
   if (isPlainObject(base)) {
@@ -50,7 +57,7 @@ export function createBaseComponent<
       Props,
       ThemeType,
       AllHTMLAttributes<Element>
-    >(base);
+    >(base as BaseConfig<Props, AllHTMLAttributes<Element>, ThemeType>);
   }
 
   return dict;
@@ -60,13 +67,16 @@ export function generateComponent<
   ThemeType = any,
   Props = Record<string, unknown>,
   Element = HTMLDivElement
->(base, setting) {
+>(
+  base: BaseConfig<Props, AllHTMLAttributes<Element>, ThemeType>,
+  setting: Settings<Element>
+) {
   return createStyledComponent<Props, ThemeType, AllHTMLAttributes<Element>>({
     ...base,
     styles: { ...base.styles, ...setting },
     attrs: { ...base.attrs, ...setting.attrs } || {},
     element: !!setting.as ? setting.as : base.element
-  } as BaseConfig<Props & Variant<ThemeType>, ThemeType, AllHTMLAttributes<Element>>);
+  } as BaseConfig<Props & Variant<ThemeType>, AllHTMLAttributes<Element>, ThemeType>);
 }
 
 /* eslint-disable @typescript-eslint/no-explicit-any */