+++ packages/usematchmedia/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @artifak/usematchmedia
 
+## 1.0.2
+
+### Patch Changes
+
+- Add window object check for useMatchMedia
+
 ## 1.0.1
 
 ### Patch Changes