+++ packages/usewindowsize/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @artifak/usewindowsize
 
+## 1.0.2
+
+### Patch Changes
+
+- Additional tests for resize event for useWindowSize
+
 ## 1.0.1
 
 ### Patch Changes